5. Always look up

+++ 0x06-pointers_arrays_strings/5-string_toupper.c
@@ -11,7 +11,7 @@
 
 char *string_toupper(char *str)
 {
-	long unsigned int i;
+	unsigned int i;
 
 	for (i = 0; i <= strlen(str); i++)
 	{
